+++ internal/storage/postgresql/migrations/209_add_changelog_v5_5_0.down.sql
@@ -1,0 +1,1 @@
+DELETE FROM changelogs WHERE version = '5.5.0';